requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.27.0
celery==5.3.6
redis==5.0.3
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
jsonschema==4.21.1
sqlalchemy==2.0.35
alembic==1.13.1
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-dotenv==1.0.1
structlog==24.1.0
sentry-sdk==1.41.0
pytest==8.1.1
pytest-asyncio==0.23.6
